# components/tenant-api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_vencil_tenant-api_internal_federation_fedpolicy.Config:
    properties:
      whitelist:
        items:
          $ref: '#/definitions/github_com_vencil_tenant-api_internal_federation_fedpolicy.WhitelistEntry'
        type: array
    type: object
  github_com_vencil_tenant-api_internal_federation_fedpolicy.Subset:
    properties:
      metrics:
        items:
          type: string
        type: array
    type: object
  github_com_vencil_tenant-api_internal_federation_fedpolicy.WhitelistEntry:
    properties:
      metric:
        type: string
    type: object
...
  internal_handler.PutFederationPolicyRequest:
    properties:
      force:
        type: boolean
      reason:
        type: string
      whitelist:
        items:
          $ref: '#/definitions/github_com_vencil_tenant-api_internal_federation_fedpolicy.WhitelistEntry'
        type: array
    type: object
...
info:
  contact: {}
  description: |-
    Multi-tenant configuration management API for the Dynamic Alerting Platform.
    Provides CRUD over per-tenant alert thresholds, RBAC-scoped writes, and async batch operations.
    Auth is enforced by the oauth2-proxy sidecar and forwarded via X-Forwarded-Email / X-Forwarded-Groups.
  license:
    name: See repository LICENSE
  title: Tenant API
  version: v2.8.0
paths:
  /api/v1/federation/policy:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vencil_tenant-api_internal_federation_fedpolicy.Config'
      summary: Get the platform federation whitelist
      tags:
      - federation
    put:
      consumes:
      - application/json
      parameters:
      - description: Whitelist + optional admission force/reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.PutFederationPolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Replace the platform federation whitelist
      tags:
      - federation
...
  /api/v1/tenants/{id}/federation:
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vencil_tenant-api_internal_federation_fedpolicy.Subset'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a tenant's effective federation metric subset
      tags:
      - federation
    put:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Metric subset
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_vencil_tenant-api_internal_federation_fedpolicy.Subset'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Replace a tenant's federation metric subset
      tags:
      - federation
...
swagger: "2.0"